Major Initial Update
After realizing, i was not account for scrabble words that can still be played because they make a word with an already played word, I am essentially redoing the main functionality for the board class - moving code from the line class to the board class.  This makes more sense in my eyes although it's gonna unfortunatley create a lot of duplicate methods that only differ on one being column checking and the other being row checking.  At this point, got a lot of testing and busy work to do.